+++ src/config/services/index.ts
@@ -1,3 +1,2 @@
 export * from './AppUnauthorizedException.service';
 export * from './token-extractor.service';
-export * from './prisma.db.service';