+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # 1.0.0 (2026-01-27)
 
-
 ### Features
 
-* **init app and setup google auth and credentials:** create nest app ([9dd5e82](https://github.com/victory-aime/my-immo-backend/commit/9dd5e826a4782d794fca9d158bc6a7fb8b45562f))
+- **init app and setup google auth and credentials:** create nest app ([9dd5e82](https://github.com/victory-aime/my-immo-backend/commit/9dd5e826a4782d794fca9d158bc6a7fb8b45562f))